net/asterisk11: Update to Asterisk 11.10.0.
RT200NE patch update too.

+++ net/asterisk11/files/patch-rt200ne-chan_sip.c
@@ -103,7 +103,7 @@
 + 	}
 + 	/* RT-200NE HACK END */
 + 
-  	snprintf(cnonce, sizeof(cnonce), "%08lx", ast_random());
+  	snprintf(cnonce, sizeof(cnonce), "%08lx", (unsigned long)ast_random());
   
   	/* Check if we have peer credentials */
 ***************
